requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
python-multipart
fastapi-mcp